Fix build errors with modern MinGW32
There is a lot of bad code in here that older versions of GCC accepted.

The bespoke `memmove` implementation is completely unnecessary. It seems
to exist either because the standard library implementation was not
optimized enough at the time, or was not portable. It's fine now, so
let's simplify and avoid the compile errors at the same time.